pyproject.toml
[project]
name = "multi-contractagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "langsmith>=0.7.14",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-docx>=1.2.0",
    "qdrant-client>=1.17.0",
]
